currency-calculator

+++ currency-calculator/src/main/java/currency/calulcator/strategies/RoundingStrategy.java
@@ -1,8 +1,10 @@
 package currency.calulcator.strategies;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Pawel on 2016-11-24.
  */
 public interface RoundingStrategy {
-     int round(float value);
+     int round(BigDecimal value);
 }
