+++ currency-calculator/src/main/java/currency/calulcator/strategies/RoundingStrategy2Up.java
@@ -1,12 +1,14 @@
 package currency.calulcator.strategies;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Pawel on 2016-11-24.
  */
 public class RoundingStrategy2Up implements RoundingStrategy {
 
     @Override
-    public int round(float value) {
+    public int round(BigDecimal value) {
         return 0;
     }
